requirements.txt
flask==3.1.1
openai==1.86.0
nltk==3.9.1
python-dotenv==1.1.0
google-generativeai==0.7.1
